2 completer bookcontroller de lapi

+++ ASP.Server/Data/Book.cs
@@ -14,7 +14,7 @@ namespace ASP.Server.Model
         public string Title { get; set; }
         public string Author { get; set; }
         public string Content { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public List<Genre> Genres { get; set; }
     }
 }
